melhorias no codigo
# Conflicts:
#	pom.xml
#	src/main/java/com/chamada_juninho/dto/RegistraAlunodto.java
#	src/main/java/com/chamada_juninho/entity/RegistraAlunos.java
#	src/main/resources/application.properties

+++ src/main/java/com/chamada_juninho/service/ChamadaService.java
@@ -1,12 +1,12 @@
 package com.chamada_juninho.service;
 
 import com.chamada_juninho.entity.RegistraAlunos;
+import com.chamada_juninho.exception.ResourceNotFoundException;
 import com.chamada_juninho.repository.RegistraAlunorepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ChamadaService {
@@ -17,9 +17,9 @@ public class ChamadaService {
         return registraAlunorepository.findAll();
     }
 
-    public Optional<RegistraAlunos> buscarPorId(Long id) {
-        return registraAlunorepository.findById(id);
-
+    public RegistraAlunos buscarPorId(Long id) {
+        return registraAlunorepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Aluno não encontrado com o ID: " + id));
     }
 
     public RegistraAlunos salvar(RegistraAlunos aluno) {
@@ -27,8 +27,7 @@ public class ChamadaService {
     }
 
     public RegistraAlunos atualizar(Long id, RegistraAlunos aluno) {
-        RegistraAlunos existingAluno = registraAlunorepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Aluno não encontrado com o ID: " + id));
+        RegistraAlunos existingAluno = buscarPorId(id); // Reutiliza o método que já lança a exceção
         existingAluno.setId(id);
         existingAluno.setNome(aluno.getNome());
         existingAluno.setTelefone(aluno.getTelefone());
@@ -37,6 +36,9 @@ public class ChamadaService {
     }
 
     public void deletar(Long id) {
+        if (!registraAlunorepository.existsById(id)) {
+            throw new ResourceNotFoundException("Aluno não encontrado com o ID: " + id);
+        }
         registraAlunorepository.deleteById(id);
     }
 }
